Refactor: StoreDetailPage 컴포넌트 분리
