+++ src/components/store/index.js
@@ -1,3 +1,6 @@
 export { default as StorePositionMap } from './StorePositionMap';
 export { default as Title } from './Title';
 export { default as DetailSide } from './DetailSide';
+export { default as Details } from './Details';
+export { default as Votes } from './Votes';
+export { default as Comments } from './Comments';
